Refactor(helm): extract chart fetching into chart_fetch.go
Move detectChartSourceType, isMutableVersion, fetchChart
  (cache wrapper), fetchChartWithoutCache, determineCacheTTL, and the
  three per-source fetchers (fetchOCIChart, fetchURLChart,
  fetchRepoChart) out of helm.go.

No behavior change. All 240 unit specs remain green.
